[REFACT] HKProviderError 파일 분리
- path: AtchI/HealthKit/Provider/Error/HKProviderError.swift
